pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gym-drones"
version = "0.3.0"
description = "Simple gym for drones RL"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24,<2.0",
    "scipy>=1.10",
    "matplotlib",
    "pyyaml",
    "pytest",
    "gymnasium",
    "stable-baselines3 @ file:./gym_drones/utils/stable-baselines3",
    "tensorboard",
    "tqdm",
    "rich",
    "race-utils @ file:./gym_drones/utils/RaceUtils",
]

[project.optional-dependencies]
dev = [
    "black==24.10.0",
    "pre-commit",
    "pytest-cov",
]

[tool.setuptools.package-data]
"gym_drones" = ["assets/**/*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["gym_drones", "gym_drones.*"]

[tool.black]
line-length = 120
extend-exclude = '''
/(
    gym_drones/utils/stable-baselines3
  | gym_drones/utils/RaceUtils
)/
'''